Extract AuthenticationStoreSync component
